remote build flow defined, patch command adapted, `-all` flag introduced, will be finished later

+++ cli/remote_build.go
@@ -3,6 +3,7 @@ package cli
 import (
 	"bytes"
 	"context"
+	"encoding/base64"
 	"encoding/json"
 	"flag"
 	"fmt"
@@ -31,54 +32,35 @@ import (
 type RemoteCmd struct {
 	target         string
 	baseCommit     string
+	commonCommit   string
 	branch         string
 	noAcceleration bool
-	foundCommit    string
-	foundTarget    string
 	patchFile      string
+	patchPath      string
+	noDeleteTemp   bool
 }
 
 func (*RemoteCmd) Name() string     { return "remotebuild" }
 func (*RemoteCmd) Synopsis() string { return "Build remotely." }
 func (*RemoteCmd) Usage() string {
-	return "Build remotely, you should login first to acquire the API_KEY for YourBase"
+	return "Build remotely using YB infrastructure"
 }
 
 func (p *RemoteCmd) SetFlags(f *flag.FlagSet) {
 	f.StringVar(&p.target, "target", "", "Repository to remote build")
 	f.StringVar(&p.baseCommit, "base-commit", "", "Base commit hash as common ancestor")
 	f.StringVar(&p.branch, "branch", "", "Branch name")
+	f.StringVar(&p.patchPath, "patch-path", "", "Path to save the patch")
 	f.BoolVar(&p.noAcceleration, "no-accel", false, "Disable accelaration")
+	f.BoolVar(&p.noDeleteTemp, "no-delete-temp", false, "Delete temp generated data")
 
-}
-
-func (p *RemoteCmd) Commit() string {
-	if p.baseCommit == "" {
-		return p.foundCommit
-	} else {
-		return p.baseCommit
-	}
-}
-
-func (p *RemoteCmd) Target() string {
-	if p.target == "" {
-		return p.foundTarget
-	} else {
-		return p.target
-	}
-}
-
-func (p *RemoteCmd) Branch() string {
-	if p.branch == "" {
-		return "master"
-	} else {
-		return p.branch
-	}
 }
 
 func (p *RemoteCmd) Execute(_ context.Context, f *flag.FlagSet, _ ...interface{}) subcommands.ExitStatus {
 
-	buildTarget := "default"
+	if p.target == "" {
+		p.target = "default"
+	}
 
 	var targetPackage Package
 
@@ -120,13 +102,13 @@ func (p *RemoteCmd) Execute(_ context.Context, f *flag.FlagSet, _ ...interface{}
 			fmt.Printf("Default build command has no steps and no targets described\n")
 		}
 	} else {
-		if _, err := manifest.BuildTarget(buildTarget); err != nil {
-			fmt.Printf("Build target %s specified but it doesn't exist!\n", buildTarget)
+		if _, err := manifest.BuildTarget(p.target); err != nil {
+			fmt.Printf("Build target %s specified but it doesn't exist!\n", p.target)
 			fmt.Printf("Valid build targets: %s\n", strings.Join(manifest.BuildTargetList(), ", "))
 		}
 	}
 
-	fmt.Printf("Remotely building '%s' ...\n", buildTarget)
+	fmt.Printf("Remotely building '%s' ...\n", p.target)
 	packagePath := targetPackage.Path
 	workRepo, err := git.PlainOpen(packagePath)
 
@@ -150,13 +132,6 @@ func (p *RemoteCmd) Execute(_ context.Context, f *flag.FlagSet, _ ...interface{}
 		for _, u := range c.URLs {
 			repoUrls = append(repoUrls, u)
 		}
-	}
-
-	// Define which commitHash to use
-	p.foundCommit, err = defineCommit(workRepo, p.baseCommit)
-	if err != nil {
-		fmt.Printf("Error finding base-commit argument: %v\n", err)
-		return subcommands.ExitFailure
 	}
 
 	project, err := fetchProject(repoUrls)
@@ -183,6 +158,7 @@ func (p *RemoteCmd) Execute(_ context.Context, f *flag.FlagSet, _ ...interface{}
 		if err = file.Close(); err != nil {
 			//Ignore
 		}
+		// We just need the unique name
 		os.Remove(patchFile)
 	}
 
@@ -191,20 +167,72 @@ func (p *RemoteCmd) Execute(_ context.Context, f *flag.FlagSet, _ ...interface{}
 		fmt.Printf("Unable to create clone dir to fetch and checkout remote repository '%v': %v\n", project.Repository, err)
 		return subcommands.ExitFailure
 	}
+	if p.noDeleteTemp {
+		defer os.RemoveAll(cloneDir)
+	}
 
-	_, err = CloneRepository(project.Repository, cloneDir, p.branch)
+	// First things first:
+	// 1. Define correct branch name
+	// 2. Try to clone the remote repository using the defined branch
+	// 3. Then try to clone it using the default 'master' branch
+	// 4. Define baseCommit (informed, validated or found)
+	// 5. If branch isn't on remote yet,
+	// 6. Define common ancestor commit
+	// 7. Generate patch file
+	// 8. Submit build!
+
+	foundBranch, err := defineBranch(workRepo, p.branch)
+
 	if err != nil {
-		fmt.Printf("Unable to clone %v: %v\n", project.Repository, err)
+		fmt.Printf("Unable to define a branch: %v\n", err)
 		return subcommands.ExitFailure
 	}
-	defer os.RemoveAll(cloneDir)
+
+	branchOnRemote := false
+	LOGGER.Debugf("Cloning repo %s into %s, using found branch '%s'", project.Repository, cloneDir, foundBranch)
+
+	clonedRepo, err := CloneRepository(project.Repository, cloneDir, foundBranch)
+	if err != nil {
+		fmt.Printf("Unable to clone %v, using branch '%v': %v\n", project.Repository, foundBranch, err)
+
+		clonedRepo, err = CloneRepository(project.Repository, cloneDir, "master")
+		if err != nil {
+			fmt.Printf("Unable to clone %v using the master branch: %v\n", project.Repository, err)
+			return subcommands.ExitFailure
+		}
+		p.branch = "master"
+	} else {
+		branchOnRemote = true
+		p.branch = foundBranch
+	}
+
+	foundCommit, err := defineCommit(workRepo, p.baseCommit)
+	if err != nil {
+		fmt.Printf("Error finding and deciding which base-commit to use: %v\n", err)
+		return subcommands.ExitFailure
+	} else {
+		fmt.Printf("Found commit: '%v'\n", foundCommit)
+		p.baseCommit = foundCommit
+	}
+
+	if !branchOnRemote {
+		targetSet := commitSet(workRepo)
+		if targetSet == nil {
+			fmt.Printf("Branch isn't on remote yet and couldn't build a commit set for comparing\n")
+			return subcommands.ExitFailure
+		}
+
+		p.commonCommit = findCommonAncestor(clonedRepo, targetSet).String()
+	} else {
+		p.commonCommit = p.baseCommit
+	}
 
 	patchCmd := &PatchCmd{
 		targetRepository: cloneDir,
 		patchFile:        patchFile,
 	}
 
-	fmt.Printf("targetRepository: %v patchFile: %v\n", patchCmd.targetRepository, patchCmd.patchFile)
+	LOGGER.Debugf("targetRepository: %v patchFile: %v\n", patchCmd.targetRepository, patchCmd.patchFile)
 
 	if patchCmd.Execute(context.Background(), f, nil) != subcommands.ExitSuccess {
 		fmt.Printf("Patch command failed\n")
@@ -212,6 +240,14 @@ func (p *RemoteCmd) Execute(_ context.Context, f *flag.FlagSet, _ ...interface{}
 	}
 
 	p.patchFile = patchFile
+
+	if p.patchPath != "" {
+		if err := savePatch(p); err != nil {
+			fmt.Printf("\nUnable to save copy of generated patch: %v\n\n", err)
+		} else {
+			fmt.Printf("Patch copy saved at: %v\n", p.patchPath)
+		}
+	}
 
 	err = submitBuild(project, p, target.Tags)
 
@@ -328,12 +364,39 @@ func postToDispatcher(path string, formData url.Values) (*http.Response, error) 
 	return res, nil
 }
 
+func defineBranch(r *git.Repository, hintBranch string) (string, error) {
+
+	ref, err := r.Head()
+	if err != nil {
+		return "", fmt.Errorf("No Head: %v\n", err)
+	}
+
+	if ref.Name().IsBranch() {
+		if hintBranch != "" {
+			if hintBranch == ref.Name().Short() {
+				fmt.Printf("Informed branch is the one used locally")
+
+				return hintBranch, nil
+
+			} else {
+				return hintBranch, fmt.Errorf("Informed branch (%v) isn't the same as the one used locally (%v)", hintBranch, ref.Name().String())
+			}
+		} else {
+			LOGGER.Debugf("Found branch reference name is %v", ref.Name().Short())
+			return ref.Name().Short(), nil
+		}
+
+	} else {
+		return "", fmt.Errorf("No branch set?")
+	}
+}
+
 func defineCommit(r *git.Repository, commit string) (string, error) {
 
 	if commit == "" {
 		ref, err := r.Head()
 		if err != nil {
-			fmt.Printf("No Head: %v\n", err)
+			return "", fmt.Errorf("No Head: %v\n", err)
 		}
 
 		return ref.Hash().String(), nil
@@ -433,6 +496,18 @@ func fetchUserEmail() (string, error) {
 	}
 }
 
+func savePatch(cmd *RemoteCmd) error {
+
+	data, _ := ioutil.ReadFile(cmd.patchFile)
+	err := ioutil.WriteFile(cmd.patchPath, data, 0644)
+
+	if err != nil {
+		return fmt.Errorf("Couldn't save a local patch file at: %v, because: %v", err)
+	}
+
+	return nil
+}
+
 func submitBuild(project *Project, cmd *RemoteCmd, tagMap map[string]string) error {
 
 	userToken, err := getUserToken()
@@ -449,14 +524,16 @@ func submitBuild(project *Project, cmd *RemoteCmd, tagMap map[string]string) err
 		return fmt.Errorf("Couldn't compress the patch file: \n    %s\n", err)
 	}
 
+	patchEncoded := base64.StdEncoding.EncodeToString(patchBuffer.Bytes())
+
 	formData := url.Values{
-		"project_id":    {strconv.Itoa(project.Id)},
-		"repository_id": {project.Repository},
-		"api_key":       {userToken},
-		"target":        {cmd.Target()},
-		"patch_data":    {patchBuffer.String()},
-		"commit":        {cmd.Commit()},
-		"branch":        {cmd.Branch()},
+		"project_id": {strconv.Itoa(project.Id)},
+		"repository": {project.Repository},
+		"api_key":    {userToken},
+		"target":     {cmd.target},
+		"patch_data": {patchEncoded},
+		"commit":     {cmd.commonCommit},
+		"branch":     {cmd.branch},
 	}
 
 	tags := make([]string, 0)
@@ -468,7 +545,11 @@ func submitBuild(project *Project, cmd *RemoteCmd, tagMap map[string]string) err
 		formData.Add("tags[]", tag)
 	}
 
-	defer os.Remove(cmd.patchFile)
+	LOGGER.Debugf("Calling backend with the following values: %v\n", formData)
+
+	if cmd.noDeleteTemp {
+		defer os.Remove(cmd.patchFile)
+	}
 
 	resp, err := postToApi("builds/cli", formData)
 	if err != nil {
@@ -484,9 +565,19 @@ func submitBuild(project *Project, cmd *RemoteCmd, tagMap map[string]string) err
 
 	response := string(body)
 
-	if strings.HasPrefix(response, "ws:") || strings.HasPrefix(response, "wss:") {
-		fmt.Printf("Streaming build output from %s\n", response)
-		conn, _, _, err := ws.DefaultDialer.Dial(context.Background(), response)
+	//Process simple response from the API
+	response = strings.ReplaceAll(response, "\"", "")
+
+	url := ""
+	if strings.Count(response, "\n") > 0 {
+		url = strings.Split(response, "\n")[0]
+	} else {
+		url = response
+	}
+
+	if strings.HasPrefix(url, "ws:") || strings.HasPrefix(url, "wss:") {
+		fmt.Printf("Streaming build output from %s\n", url)
+		conn, _, _, err := ws.DefaultDialer.Dial(context.Background(), url)
 		if err != nil {
 			return fmt.Errorf("Can not connect: %v\n", err)
 		} else {
@@ -514,8 +605,7 @@ func submitBuild(project *Project, cmd *RemoteCmd, tagMap map[string]string) err
 			}
 		}
 	} else {
-		fmt.Println("Wrong response:")
-		fmt.Println(response)
+		return fmt.Errorf("Unable to stream build output!")
 	}
 
 	return nil
